% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceleration_vector.R
\name{accelerometer_data}
\alias{accelerometer_data}
\title{Import of accelerometer data and change class on time variable}
\usage{
accelerometer_data(data)
}
\arguments{
\item{data}{The name of the file which the data are to be read from.}
}
\value{
A dataframe with the calculated "acceleration_vector"
}
\description{
Import of accelerometer data and change class on time variable
}
