% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceleration_vector.R
\name{euclidean_calculation}
\alias{euclidean_calculation}
\title{Calculation of euclidean norm}
\usage{
euclidean_calculation(data)
}
\arguments{
\item{data}{Data frame containing index, time, three columns containing x,y,and z}
}
\value{
A euclidean_norm, and time vector
}
\description{
Calculation of euclidean norm
}
\examples{
# example code
 #Upload the practice data
 data("accelerometer_data")

 #Run the function
 euclidean_calculation <- function(accelerometer_data)

}
