% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_acceleration.R
\name{plot_accelerometer}
\alias{plot_accelerometer}
\title{Plot accelerometer data}
\usage{
plot_accelerometer(data, variable = "euclidean_norm", option = "mean")
}
\arguments{
\item{data}{Data.frame containing variable "mean" and "binary" which you get from running
window_threshold_test() or window_function_test_mean().}

\item{variable}{Could be any axis "x", "y", "z", or "euclidean_norm". Default is "euclidean_norm"}

\item{option}{Specify if you want to plot "mean", "binary", or "mean_binary"}
}
\value{
A plot object
}
\description{
Plot accelerometer data
}
