% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_acceleration.R
\name{plot_accelerometer_data}
\alias{plot_accelerometer_data}
\title{Plot accelerometer data depending on method (mean or peak averages for values over threshold)}
\usage{
plot_accelerometer_data(
  data,
  calculation = "mean",
  column = "euclidean_norm",
  option = "mean"
)
}
\arguments{
\item{data}{Data.frame containing containing euclidean_norm.}

\item{calculation}{Supply "mean" if you want to calculate mean if acceleration is above threshold
or "peak" if you want to supply largest differences (positive or negative)}

\item{column}{Which column to run function on. Supply "x", "y", "z", or "euclidean_norm"}

\item{option}{Specify if you want to plot "mean", "binary", or both "mean_binary"}
}
\value{
A plot object
}
\description{
Plot accelerometer data depending on method (mean or peak averages for values over threshold)
}
