% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceleration_window.R
\name{window_function_test_mean}
\alias{window_function_test_mean}
\title{Calculates the mean value of a sliding window}
\usage{
window_function_test_mean(data, column, window, threshold)
}
\arguments{
\item{data}{Data frame containing index, time, three columns containing x,y,z, and the euclidean_norm}

\item{column}{Can be "x", "y", "z" or "euclidean_norm". Default is "x"}

\item{window}{How large you want the indexed window to be. Default is 1000}

\item{threshold}{A numeric threshold where changes in acceleration exceeding the threshold is manipulated. Default is 0.5}
}
\value{
Two columns "mean" (whenever acceleration exceed threshold mean for window is applied) and "binary" ("+1" or "-1" if acceleration is positive or negative)
}
\description{
Calculates the mean value of a sliding window
}
\examples{
# example code
 #Upload the data
 data("accelerometer_data")

 #euclidean_norm
 accelerometer_data <- euclidean_calculation(accelerometer_data)

 #Run function
 accelerometer_data <- window_threshold_test(accelerometer_data)

}
