% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceleration_window.R
\name{window_threshold_test}
\alias{window_threshold_test}
\title{Extracts and applies window min or max}
\usage{
window_threshold_test(
  data,
  column = "euclidean_norm",
  window = 1000,
  threshold = 0.5
)
}
\arguments{
\item{data}{Data frame containing index, time, three columns containing x,y,z, and the euclidean_norm}

\item{column}{Can be "x", "y", "z" or "euclidean_norm". Default is "x"}

\item{window}{How large you want the indexed window to be. Default is 1000}

\item{threshold}{A numeric threshold where changes in acceleration exceeding the threshold is manipulated. Default is 0.5}
}
\value{
New columns "mean" and "binary"
}
\description{
Extracts and applies window min or max
}
\examples{
# example code
 #Upload the data
 data("accelerometer_data")

 #euclidean_norm
 accelerometer_data <- euclidean_calculation(accelerometer_data)

 #Run function
 accelerometer_data <- window_threshold_test(accelerometer_data)

}
